TXT functionality
added logic to take in TXT files and parse through and get lat,lon given specified formatting
